Put recommended modules into TEST_DEPENDS
Because there's a good chance we actually want them for tests.

Almost as likely is that they are actual depends, not just recommended
so not having them will cause breakage that doesn't show up when
testing.  I assume it is better to assume people know what they are
doing in this particular instance.